Рефакторит trip-info-view
